+++ src/mypy_upgrade/filter.py
@@ -1,3 +1,4 @@
+"""This module defines functions for filtering mypy type checking errors."""
 # remove when dropping Python 3.7-3.9 support
 from __future__ import annotations
 
@@ -33,6 +34,10 @@ def get_module_paths(modules: list[str]) -> list[pathlib.Path | None]:
                 module_path = pathlib.Path(loader.get_filename(module))
                 if loader.is_package(module):
                     module_path = module_path.parent
+            elif spec.origin == "frozen":
+                module_path = pathlib.Path(spec.loader_state.filename)
+            elif spec.origin == "built-in":
+                pass  # ? Maybe should raise a warning
             else:
                 msg = "Uncountered an unsupported module type."
                 raise NotImplementedError(msg)